requirements.txt
# Windows Developer Platform Agent - Dependencies
# All paths use pathlib; platform-agnostic Python.

# Core
httpx>=0.25.0
pydantic-settings>=2.0.0

# LLM / Agent
# (OpenRouter, OpenAI, Ollama use same chat API)

# Integrations
PyGithub>=2.1.0
slack-sdk>=3.23.0
jira>=3.5.0
atlassian-python-api>=3.34.0
python-jenkins>=1.7.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0

# Database
SQLAlchemy>=2.0.0

# Web
FastAPI>=0.100.0
uvicorn>=0.22.0

# Workflows
PyYAML>=6.0.0

# CLI
rich>=13.0.0
